channels-bot/requirements.txt
aiogram
aiokafka
python-dotenv
minio
redis>=4.2.0